Extracted a public interface for running the generator
Extracted a public interface for running the generator
outside of a Gradle task. Gradle doesn't let you create
tasks programmatically.

+++ macfiles-native/src/test/kotlin/garden/ephemeral/macfiles/native/NativeTest.kt
@@ -100,7 +100,7 @@ class NativeTest {
             .isInstanceOf(List::class)
             .endsWith("macfiles-native", "build.gradle.kts")
         assertThat(toc[BookmarkKeys.kBookmarkCNIDPath]).isNotNull().isInstanceOf(List::class)
-            .index(0).isNotNull().isInstanceOf(UInt::class)
+            .index(0).isNotNull().isInstanceOf(Long::class)
         assertThat(toc[BookmarkKeys.kBookmarkFileCreationDate]).isNotNull().isInstanceOf(Instant::class)
         assertThat(toc[BookmarkKeys.kBookmarkFileProperties]).isEqualTo(
             Blob.fromHexDump("""
@@ -144,7 +144,7 @@ class NativeTest {
             .isInstanceOf(List::class)
             .endsWith("macfiles-native", "build")
         assertThat(toc[BookmarkKeys.kBookmarkCNIDPath]).isNotNull().isInstanceOf(List::class)
-            .index(0).isNotNull().isInstanceOf(UInt::class)
+            .index(0).isNotNull().isInstanceOf(Long::class)
         assertThat(toc[BookmarkKeys.kBookmarkFileCreationDate]).isNotNull().isInstanceOf(Instant::class)
         assertThat(toc[BookmarkKeys.kBookmarkFileProperties]).isEqualTo(
             Blob.fromHexDump("""
